+++ NEWS.rst
@@ -7,6 +7,8 @@ of every change, see the Git log.
 Latest
 ------
 * tbd
+* Patch: If ssh_output_file used, then append the shellexit line to the
+  output file. This is useful if the SSH output is truncated from some reason.
 * Patch: The configure step should fail when the specified version of msvc
   is not found.